+++ components/Header/Header.jsx
@@ -4,6 +4,7 @@ import React, { useState, useEffect } from "react";
 import Link from 'next/link';
 import { useRouter } from 'next/navigation';
 import styles from "./styles.module.css";
+import { doLogout } from "./utils";
 
 
 const Header = () => {
@@ -19,37 +20,7 @@ const Header = () => {
         }
     }, []);
 
-    const handleLogout = async () => {
-        if (typeof window !== "undefined") {
-          const refreshToken = localStorage.getItem("refresh-token");
-    
-          // Llamada para eliminar (blacklistear) el refresh token
-          if (refreshToken) {
-            try {
-              await fetch("http://127.0.0.1:8000/api/token/refresh/", {
-                method: "POST",
-                headers: {
-                  "Content-Type": "application/json",
-                },
-                body: JSON.stringify({ refresh: refreshToken }),
-              });
-            } catch (error) {
-              console.error("Error al invalidar el refresh token:", error);
-            }
-          }
-    
-          // Limpiar localStorage y estado
-          localStorage.removeItem("token-jwt");
-          localStorage.removeItem("refresh-token");
-          localStorage.removeItem("userName");
-    
-          setToken(null);
-          setUserName(null);
-    
-          // Redirigir al inicio de sesión
-          router.push("/inicio");
-        }
-    };
+    const handleLogout = async () => doLogout(setToken, setUserName, router);
     
     return (
         <header className={styles.header}>
